test: fold the cancel regression tests into the existing suites
Review feedback: a separate file per test is not worth it. The idempotency
case now reuses the package's validKey constant and do helper instead of
duplicating them.
